requirements.txt
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
PyJWT==2.2.0